Added bitvector.ml, sample build script, common c
bitvector.ml compiles four modules for the width of bitvector computation. It is
not scripted by the types of widths, but since this doesn't change much it's
approprate. BV8, BV16, BV32, and BV64 are sets of external functions as a result
of bv.c scripted for each bit-width. A build script for this process is included
until a proper build system is choosen --oasis is not going to be appropriate as
the previous changeset notes.

phyloc.[ch] has common parts of phylo-caml c functions. This includes ways in
dealing with options on the C-Side.