Update for 4.01.0ocaml; alphabet+mlmodel updates
Create a better interface for the alphabet module. We don't require the codes to
be specified by the user on of_list function. This was uncesseccary and would
lead to more stringent verification of the alphabet.

Remove printing and categorization functions from mlmodel.

update $ to be @@ as in the stdlib