Remove the diagnosis modules (1); Rename Pretty Printer Module (2)
(1) Removing the diagnosis module since this is an operation for a particular
application to decide. This relieves a lot of engineering stress to create
something general enough to solve all problems --impossible. Examples will be
provided in the app/ directory for particular diagnosis methods for ease of use
though.

(2) Ppp module named to Ppf (Pretty-Printer Formating). Contains utility
functions for pretty-printing types -- array, matrix, table -- instead of the
container of all pp functions -- would be incomplete due to functors anyways.